+++ CHANGES.rst
@@ -5,6 +5,7 @@ mqttwarn changelog
 
 in progress
 ===========
+- [udf] Unlock JavaScript for user-defined functions. Thanks, @extremeheat.
 
 
 2023-10-15 0.35.0